% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transit.R
\name{get_lines}
\alias{get_lines}
\title{Lines}
\usage{
get_lines(operator_id)
}
\arguments{
\item{operator_id}{filter lines by operator ID}
}
\value{
Dataframe of transit lines
}
\description{
Lines are routes covered by transit operators.
}
\examples{
get_lines(operator_id='AC')
}
