% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transit.R
\name{get_operators}
\alias{get_operators}
\title{Operators}
\usage{
get_operators()
}
\value{
Dataframe of transit operators in the Bay Area
}
\description{
Operators provide public transport services.
}
\examples{
get_operators()
}
