% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stop_predictions_for_operators.R
\name{get_realtime_predictions_for_operator}
\alias{get_realtime_predictions_for_operator}
\title{Realtime Predictions for all Stops for an Operator}
\usage{
get_realtime_predictions_for_operator(operator_id)
}
\arguments{
\item{operator_id}{Operator ID to filter by}
}
\value{
Dataframe of all predictions for stops covered by a particular operator up to 1000 stops
}
\description{
Realtime Predictions for all Stops for an Operator
}
\examples{
get_realtime_predictions_for_operator(operator_id='AC')
}
