% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realtime_transit.R
\name{get_realtime_predictions_for_stop}
\alias{get_realtime_predictions_for_stop}
\title{Real Time Predictions at a Stop}
\usage{
get_realtime_predictions_for_stop(operator_id, stop_id)
}
\arguments{
\item{operator_id}{ID of transit agency}

\item{stop_id}{StopCode referring to a stop}
}
\value{
Single row DataFrame with prediction for the specific stop
}
\description{
Get real time arrival prediction for a specific stop ID
}
\examples{
get_realtime_predictions_for_stop(operator_id='AC', stop_id=53326)
}
