% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transit.R
\name{get_stopPlaces}
\alias{get_stopPlaces}
\title{StopPlace}
\usage{
get_stopPlaces(operator_id, stop_id = NA)
}
\arguments{
\item{operator_id}{Operator ID is used to filter the request}

\item{stop_id}{ID identifying the stop}
}
\value{
Dataframe of stops, columns are characteristics of stops
}
\description{
Description of each stop, including accessibility, address, and parking features
}
\examples{
get_stopPlaces(operator_id='AC')
get_stopPlaces(operator_id='AC', stop_id=58772)
}
