% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transit.R
\name{get_stops}
\alias{get_stops}
\title{Stop}
\usage{
get_stops(operator_id)
}
\arguments{
\item{operator_id}{ID of the transit operator to filter the stops}
}
\value{
Dataframe of ScheduledStopPlace
}
\description{
Stop provides the locations of stops for transit lines. More details about the physical characteristics of 
of each stop can be found from StopPlace 
The stop API provides ScheduledStopPoint which describes individual stops 
as well as StopAreas which describes the parent heirarchy of stops.
}
\examples{
get_stops(operator_id='SF')
get_stops(operator_id='AC')
}
