% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traffic.R
\name{traffic_events}
\alias{traffic_events}
\title{Retrieve a list of traffic events from today.}
\usage{
traffic_events(status = "ACTIVE", headline = NA, event_type = NA,
  severity = NA, updated = NA, schedule = NA, limit = 10000)
}
\arguments{
\item{status}{Default 'ALL'.}

\item{headline}{Summary of event.}

\item{event_type}{Type of event such as CONSTRUCTION, SPECIAL_EVENT}

\item{severity}{level of severity of event. Note this is usually UNKNOWN}

\item{updated}{UTC timestamp of event updated on or after.}

\item{schedule}{UTC timestamp of times event is active}

\item{limit}{Limiting number of events to return}
}
\value{
Dataframe of events filtered to parameters passed
}
\description{
Retrieve a list of traffic events from today.
}
\examples{
traffic_events()
traffic_events(status='ACTIVE')
traffic_events(event_type='CONSTRUCTION')
}
